+++ src/ExportJobs/apiQuery.js
@@ -1,0 +1,15 @@
+export const useExportJobsQuery = () => {
+  return {
+    isLoading: false,
+    exportJobs: [
+      {
+        id: '1',
+        name: 'Job1',
+      },
+      {
+        id: '2',
+        name: 'Job2',
+      },
+    ],
+  };
+};