+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 ### Stories
 * Project Setup: ui-export-manager. Refs UIEXPMGR-1.
+* Export manager layout. Refs UIEXPMGR-2.